+++ source/extensions/omni.isaac.sensor/python/scripts/camera_view.py
@@ -11,6 +11,7 @@ from typing import Callable, List, Optional, Sequence, Tuple, Union
 import carb
 import numpy as np
 import omni.replicator.core as rep
+import omni.usd
 import torch
 import warp as wp
 from omni.isaac.core.prims.xform_prim_view import XFormPrimView
@@ -28,6 +29,40 @@ W_U_TRANSFORM = np.array([[0, 0, -1, 0], [-1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, 
 
 # from World camera convention to USD camera convention
 U_W_TRANSFORM = np.array([[0, -1, 0, 0], [0, 0, 1, 0], [-1, 0, 0, 0], [0, 0, 0, 1]])
+
+
+@wp.kernel
+def reshape_tiled_image(
+    tiled_image_buffer: wp.array(dtype=float),
+    batched_image: wp.array(dtype=float, ndim=4),
+    image_height: int,
+    image_width: int,
+    num_channels: int,
+    num_tiles_x: int,
+    offset: int,
+):
+    """Reshape a tiled image (height*width*num_channels*num_cameras,) to a batch of images (num_cameras, height, width, num_channels).
+
+    Args:
+        tiled_image_buffer: The input image buffer. Shape is ((height*width*num_channels*num_cameras,).
+        batched_image: The output image. Shape is (num_cameras, height, width, num_channels).
+        image_width: The width of the image.
+        image_height: The height of the image.
+        num_channels: The number of channels in the image.
+        num_tiles_x: The number of tiles in x direction.
+        offset: The offset in the image buffer. This is used when multiple image types are concatenated in the buffer.
+    """
+    camera_id, height_id, width_id = wp.tid()
+    tile_x_id = camera_id % num_tiles_x
+    tile_y_id = camera_id // num_tiles_x
+    pixel_start = (
+        offset
+        + num_channels * num_tiles_x * image_width * (image_height * tile_y_id + height_id)
+        + num_channels * tile_x_id * image_width
+        + num_channels * width_id
+    )
+    for i in range(num_channels):
+        batched_image[camera_id, height_id, width_id, i] = tiled_image_buffer[pixel_start + i]
 
 
 class CameraView(XFormPrimView):
@@ -67,9 +102,7 @@ class CameraView(XFormPrimView):
         self.camera_resolution = camera_resolution
 
         self._tiled_render_product = None
-        self._tiled_rgb_annotator = None
         self._tiled_annotator = None
-        self._tiled_depth_annotator = None
         self._tiled_sensor = None
 
         self._setup_tiled_sensor()
@@ -82,16 +115,9 @@ class CameraView(XFormPrimView):
         if self._tiled_render_product is not None:
             # detach annotators from render product \
             self._tiled_annotator.detach([self._tiled_render_product.path])
-            self._tiled_rgb_annotator.detach([self._tiled_render_product.path])
-            self._tiled_depth_annotator.detach([self._tiled_render_product.path])
 
             # delete tiled render products
             self._tiled_render_product.destroy()
-
-        if self._tiled_sensor is not None:
-            stage = omni.usd.get_context().get_stage()
-            for name, prim_path in self._tiled_sensor.get_input_prims().items():
-                stage.RemovePrim(prim_path)
 
     def _get_tiled_resolution(self, num_cameras, resolution) -> Tuple[int, int]:
         """Calculate the resolution for the tiled sensor based on the number of cameras and individual camera resolution.
@@ -129,8 +155,6 @@ class CameraView(XFormPrimView):
 
         self._tiled_annotator = rep.AnnotatorRegistry.get_annotator("RtxSensorGpu")
         self._tiled_annotator.attach(self._tiled_render_product)
-
-        rep.orchestrator.step()
 
         return
 
@@ -419,65 +443,97 @@ class CameraView(XFormPrimView):
 
         return XFormPrimView.set_local_poses(self, positions, orientations, indices)
 
-    def get_rgba(self, device="cpu") -> np.ndarray | torch.Tensor:
-        """Fetch the RGBA data for all cameras.
+    def get_rgb_tiled(self, out=None, device="cpu") -> np.ndarray | torch.Tensor:
+        """Fetch the RGB data for all cameras as a single tiled image.
 
         Args:
             device (str, optional): The device to return the data on ("cpu" or "cuda"). Defaults to "cpu".
+            out (np.ndarray | torch.Tensor, optional): Pre-allocated array or tensor to fill with the RGBA data.
 
         Returns:
             np.ndarray | torch.Tensor: containing the RGBA data for each camera. Depth channel is excluded if present.
         """
         if "rgb" not in self.output_annotators:
-            if device == "cuda":
-                rgb = torch.zeros((*self.tiled_resolution, 3), dtype=torch.uint8)
-            else:  # device == "cpu"
-                rgb = np.zeros((*self.tiled_resolution, 3), dtype=np.uint8)
+            if out is None:
+                if device == "cuda":
+                    out = torch.zeros((*self.tiled_resolution, 3), device=device, dtype=torch.float32)
+                else:  # device == "cpu"
+                    out = np.zeros((*self.tiled_resolution, 3), dtype=np.float32)
             print("Warning: RGB data is not available. Please enable RGB annotator when creating CameraView object.")
-        elif "depth" in self.output_annotators:  # and "rgb" in self.output_annotators
+        else:
             linear_sensor_data = self._tiled_annotator.get_data(device=device)
             if isinstance(linear_sensor_data, wp.types.array):
-                linear_sensor_data = wp.to_torch(linear_sensor_data)
-            rgb_idx = len(linear_sensor_data) * 3 // 4
-            rgb = linear_sensor_data[:rgb_idx].reshape(*self.tiled_resolution, 3) * 255
-        else:  # only "rgb" in self.output_annotators
-            linear_sensor_data = self._tiled_annotator.get_data(device=device)
-            if isinstance(linear_sensor_data, wp.types.array):
-                linear_sensor_data = wp.to_torch(linear_sensor_data)
-            rgb = linear_sensor_data.reshape(*self.tiled_resolution, 3) * 255
+                linear_sensor_data = wp.to_torch(linear_sensor_data).to(device)
 
-        return rgb
+            # If depth data is present, get the index where RGB data ends, otherwise use all the data as RGB data
+            if "depth" in self.output_annotators:
+                rgb_idx = len(linear_sensor_data) * 3 // 4
+                rgb_data = linear_sensor_data[:rgb_idx].reshape(*self.tiled_resolution, 3)
+            else:
+                rgb_data = linear_sensor_data.reshape(*self.tiled_resolution, 3)
 
-    def get_depth(self, device="cpu") -> np.ndarray | torch.Tensor:
-        """Fetch the depth data for all cameras.
+            # Check if the data should be copied to the pre-allocated memory
+            if out is not None:
+                if isinstance(out, np.ndarray):
+                    if isinstance(rgb_data, torch.Tensor):
+                        out[:] = rgb_data.cpu().numpy()
+                    else:
+                        out[:] = rgb_data
+                elif isinstance(out, torch.Tensor):
+                    if isinstance(rgb_data, np.ndarray):
+                        rgb_data = torch.from_numpy(rgb_data).to(device)
+                    out[:] = rgb_data.to(out.device)
+            else:
+                out = rgb_data
+
+        return out
+
+    def get_depth_tiled(self, out=None, device="cpu") -> np.ndarray | torch.Tensor:
+        """Fetch the depth data for all cameras as a single tiled image.
 
         Args:
             device (str, optional): The device to return the data on ("cpu" or "cuda"). Defaults to "cpu".
+            out (np.ndarray | torch.Tensor, optional): Pre-allocated array or tensor to fill with the depth data.
 
         Returns:
-            np.ndarray: containing the depth data for each camera.
+            np.ndarray | torch.Tensor: containing the depth data for each camera.
         """
         if "depth" not in self.output_annotators:
-            if device == "cuda":
-                depth = torch.zeros(self.tiled_resolution, dtype=torch.float32)
-            else:  # device == "cpu"
-                depth = np.zeros(self.tiled_resolution, dtype=np.float32)
+            if out is None:
+                if device == "cuda":
+                    out = torch.zeros(self.tiled_resolution, device=device, dtype=torch.float32)
+                else:  # device = "cpu"
+                    out = np.zeros(self.tiled_resolution, dtype=np.float32)
             print(
                 "Warning: Depth data is not available. Please enable depth annotator when creating CameraView object."
             )
-        elif "rgb" in self.output_annotators:  # and "depth" in self.output_annotators
+        else:
             linear_sensor_data = self._tiled_annotator.get_data(device=device)
             if isinstance(linear_sensor_data, wp.types.array):
-                linear_sensor_data = wp.to_torch(linear_sensor_data)
-            rgb_idx = len(linear_sensor_data) * 3 // 4
-            depth = linear_sensor_data[rgb_idx:].reshape(*self.tiled_resolution) * 255
-        else:  # only "depth" in self.output_annotators
-            linear_sensor_data = self._tiled_annotator.get_data(device=device)
-            if isinstance(linear_sensor_data, wp.types.array):
-                linear_sensor_data = wp.to_torch(linear_sensor_data)
-            depth = self._tiled_annotator.get_data().reshape(*self.tiled_resolution) * 255
+                linear_sensor_data = wp.to_torch(linear_sensor_data).to(device)
 
-        return depth
+            # If RGB data is present, depth data starts after the RGB index, otherwise the data only includes depth
+            if "rgb" in self.output_annotators:
+                rgb_idx = len(linear_sensor_data) * 3 // 4
+                depth_data = linear_sensor_data[rgb_idx:].reshape(*self.tiled_resolution)
+            else:
+                depth_data = linear_sensor_data.reshape(*self.tiled_resolution)
+
+            # Check if the data should be copied to the pre-allocated memory
+            if out is not None:
+                if isinstance(out, np.ndarray):
+                    if isinstance(depth_data, torch.Tensor):
+                        out[:] = depth_data.cpu().numpy()
+                    else:
+                        out[:] = depth_data
+                elif isinstance(out, torch.Tensor):
+                    if isinstance(depth_data, np.ndarray):
+                        depth_data = torch.from_numpy(depth_data).to(device)
+                    out[:] = depth_data.to(out.device)
+            else:
+                out = depth_data
+
+        return out
 
     def get_data(self, device="cpu") -> np.ndarray | wp.types.array:
         """Get the tiled annotator data directly.
@@ -488,7 +544,87 @@ class CameraView(XFormPrimView):
         Returns:
             np.ndarray | wp.types.array: The data from the tiled annotator.
         """
-        return self._tiled_annotator.get_data(device)
+        return self._tiled_annotator.get_data(device=device)
+
+    def get_rgb(self, out=None) -> torch.Tensor:
+        """Get the RGB data for all cameras as a batch of images (num_cameras, height, width, num_channels=3).
+
+        Returns:
+            torch.Tensor: containing the RGB data for each camera.
+            Shape is (num_cameras, height, width, num_channels=3) with type torch.float32.
+
+        """
+        # Check if the data should be copied to the pre-allocated memory
+        if out is None:
+            out_shape = (len(self.prims), *self.camera_resolution, 3)
+            out = torch.zeros(out_shape, device="cuda", dtype=torch.float32).contiguous()
+
+        # Get the linear sensor data from the tiled annotator and (if needed) slice it to get only the RGB data
+        linear_sensor_data = self._tiled_annotator.get_data(device="cuda")
+        if "depth" in self.output_annotators:
+            rgb_offset = len(linear_sensor_data) * 3 // 4
+            linear_sensor_data = linear_sensor_data[:rgb_offset]
+
+        # Use a warp kernel to convert the linear sensor data to a batch of images
+        img_height, img_width = self.camera_resolution
+        num_channels = 3
+        num_tiles_x = self.tiled_resolution[0] // img_width
+        offset = 0  # always 0 since we sliced the data above (if depth data is present)
+        wp.launch(
+            kernel=reshape_tiled_image,
+            dim=(len(self.prims), *self.camera_resolution),
+            inputs=[
+                linear_sensor_data,
+                wp.from_torch(out),
+                img_height,
+                img_width,
+                num_channels,
+                num_tiles_x,
+                offset,
+            ],
+            device="cuda",
+        )
+        return out
+
+    def get_depth(self, out=None) -> torch.Tensor:
+        """Get the depth data for all cameras as a batch of images (num_cameras, height, width, num_channels=1).
+
+        Returns:
+            torch.Tensor: containing the depth data for each camera.
+            Shape is (num_cameras, height, width, num_channels=1) with type torch.float32.
+
+        """
+        # Check if the data should be copied to the pre-allocated memory
+        if out is None:
+            out_shape = (len(self.prims), *self.camera_resolution, 1)
+            out = torch.zeros(out_shape, device="cuda", dtype=torch.float32).contiguous()
+
+        # Get the linear sensor data from the tiled annotator and (if needed) slice it to get only the depth data
+        linear_sensor_data = self._tiled_annotator.get_data(device="cuda")
+        if "rgb" in self.output_annotators:
+            depth_offset = len(linear_sensor_data) * 3 // 4
+            linear_sensor_data = linear_sensor_data[depth_offset:]
+
+        # Use a warp kernel to convert the linear sensor data to a batch of images
+        img_height, img_width = self.camera_resolution
+        num_channels = 1
+        num_tiles_x = self.tiled_resolution[0] // img_width
+        offset = 0  # always 0 since we sliced the data above (if RGB data is present)
+        wp.launch(
+            kernel=reshape_tiled_image,
+            dim=(len(self.prims), *self.camera_resolution),
+            inputs=[
+                linear_sensor_data,
+                wp.from_torch(out),
+                img_height,
+                img_width,
+                num_channels,
+                num_tiles_x,
+                offset,
+            ],
+            device="cuda",
+        )
+        return out
 
     def get_focal_lengths(
         self, indices: Optional[Union[np.ndarray, list, torch.Tensor, wp.array]] = None